Implement rtc; change hour & minute vals to bytes

+++ first_pass_v01/rtc.h
@@ -7,9 +7,9 @@ namespace rtc {
   };
 
   void setup();
-  void apply_time_delta(int hour, int minute); // Resets seconds if minute != 0
+  void apply_time_delta(byte hour, byte minute); // Resets seconds if minute != 0
   void set_hour_mode(HourMode mode);
-  void get_time(int* hour, int* minute, bool* isTwentyFourHour);
+  void get_time(byte* hour, byte* minute, bool* isTwentyFourHour);
   bool has_minute_passed();
   void reset_minute_passed();
 }
